requirements.txt
python==3.7.2
hydra-core==1.0.3
finrl==0.3.4
mbrl==0.1.4
termcolor>=1.1.0
tensorboard>=2.4.0
imageio>=2.9.0
numpy>=1.19.1
matplotlib>=3.3.1
gym==0.17.2
jupyter>=1.0.0
pytest>=6.0.1
sk-video>=1.1.10
sphinx>=3.3.1
nbsphinx>=0.8.0
sphinx-rtd-theme>=0.5.0
flake8>=3.8.4
mypy>=0.902
black>=21.4b2
pytest>=6.0.1
types-pyyaml>=0.1.6
types-termcolor>=0.1.0
git+https://github.com/quantopian/pyfolio
stable_baselines3
yfinance
stockstats
statsmodels
torch